move convert_coordinates from ome_zarr.py to utils.py
